Update README.md
- In “ Creating and Defining the Smart Contract  ”:
a)	Input Validation: The code already checks that the input fields title and url are not empty strings. However, it could further validate the URL to check that it is a valid URL, such as by using a regular expression pattern matching.
b)	Gas Optimization: The current implementation uses an increasing integer contentCount to index content in the mapping, which can cause the function to consume more gas as the number of contents grows. A better alternative is to use a cryptographic hash function to generate unique content IDs.
c)	Added error handling: If any of these checks fail, the function will revert with an error message. These error checks help ensure that the smart contract works as intended and prevent malicious or accidental misuse of the contract.

- Added relevant comments to the code as they provide a brief overview of what the contract does and explain the purpose of each function and variable.
- Applied code indentation so that it looks neat and aligned.
- Added relevant links to the source.
- Did paraphrasing.
- Added punctuations wherever, required.
- Corrected and spelling mistakes grammatical errors 
- Made code snippets.
- Formatted the entire document by adding/removing ample spaces between paragraphs so that it looks well-structured.